Renamed util to protocol
