+++ org-code-javabuilder/lib/build.gradle
@@ -27,7 +27,7 @@ dependencies {
     implementation group: 'com.amazonaws', name: 'aws-java-sdk-sqs', version: '1.11.978'
     // https://mvnrepository.com/artifact/org.json/json
     implementation group: 'org.json', name: 'json', version: '20210307'
-    implementation project(':util')
+    implementation project(':protocol')
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.0'
     // https://mvnrepository.com/artifact/org.mockito/mockito-core
     testImplementation group: 'org.mockito', name: 'mockito-core', version: '3.9.0'
