+++ org-code-javabuilder/lib/src/main/java/dev/javabuilder/LocalOutputAdapter.java
@@ -2,7 +2,7 @@ package dev.javabuilder;
 
 import java.io.PrintStream;
 import org.code.javabuilder.OutputAdapter;
-import org.code.util.ClientMessage;
+import org.code.protocol.ClientMessage;
 
 /** Intended for local testing only. Passes output to the provided PrintStream */
 public class LocalOutputAdapter implements OutputAdapter {
