+++ org-code-javabuilder/lib/src/main/java/org/code/javabuilder/SystemOutMessage.java
@@ -1,7 +1,7 @@
 package org.code.javabuilder;
 
-import org.code.util.ClientMessage;
-import org.code.util.ClientMessageType;
+import org.code.protocol.ClientMessage;
+import org.code.protocol.ClientMessageType;
 
 /** A message directed to the client's terminal. Equivalent to System.out.print. */
 public class SystemOutMessage extends ClientMessage {
