+++ org-code-javabuilder/lib/src/main/java/org/code/javabuilder/UserInitiatedExceptionMessage.java
@@ -1,8 +1,8 @@
 package org.code.javabuilder;
 
 import java.util.HashMap;
-import org.code.util.ClientMessage;
-import org.code.util.ClientMessageType;
+import org.code.protocol.ClientMessage;
+import org.code.protocol.ClientMessageType;
 
 /** An error message directed to the user, caused by the user. */
 public class UserInitiatedExceptionMessage extends ClientMessage {
