+++ org-code-javabuilder/lib/src/test/java/org/code/javabuilder/SystemOutMessageTest.java
@@ -2,7 +2,7 @@ package org.code.javabuilder;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import org.code.util.ClientMessage;
+import org.code.protocol.ClientMessage;
 import org.junit.jupiter.api.Test;
 
 public class SystemOutMessageTest {
