+++ org-code-javabuilder/settings.gradle
@@ -10,4 +10,4 @@
 rootProject.name = 'org-code-javabuilder'
 include('lib')
 include('neighborhood')
-include('util')
+include('protocol')
